Also init faro in arbeidsgiver layout
And refactor instrumentation setup a bit
