+++ src/instrumentation/faro.ts
@@ -1,5 +1,3 @@
-"use client";
-
 import {
   Faro,
   getWebInstrumentations,
